requirements.txt
aiogram==3.4.1
python-dotenv
SQLAlchemy
asyncpg
redis
